ip/torrent: fix webseed
Commit 9f755671fb6f8250da9c16a9abebbf975b8bc571 broke
webseeding with the last block.

The haveiece() call at the end was because the inner
is not calling havepiece() on the last block as it
does not take the piece length into account.

Now, instead, fix the inner loop, making the code
more setright foward so we call havepiece() on the
last block.