ip/torrent: fix webseed
Commit a73eaa7010b6a13ae5bdfa189c0f1a84c293e2ae broke
webseeding with the last block.

The haveiece() call at the end was because the inner
is not calling havepiece() on the last block as it
does not take the piece length into account.

Now, instead, fix the inner loop, making the code
more setright foward so we call havepiece() on the
last block.