nusb/*: improve usbd to handle transaction translater properties correctly
In an upcoming commit, the interface for how to create hubs
and how to update endpoint parameters is going to change.

Device/endpoint properties should not be modified while
the data file is open (device is being used).

This also applies to control endpoint when changing
packet size.

The motivation here is to clean up the xhci driver
and not do these stupid hacks like parsing control
messages. It is easier to just have the hci drivers
apply everything at open time and being guaranteed
that properties do not change under them.

For this we need to make sure to only do these devctl's
while the data file is not open.

For hubs, the command changes and some parameters.

Primarily the number of ports (required for xhci) which
will let devusb do some error checking and the
USB2.0 -> USB1.1 transaction translator properties.

In usb/lib, the "isroot" property is redundant and is
replaced by depth < 0.

For usb3.0 hub descriptor, the led indicator fields are
different from usb2.0 descriptors.

Rest is pikeshedding.