ip/ipconfig: unconditionally refresh when adding addresses manually (thanks chilledfrogs)
The issue is that ndb/cs refuses to serve ipv6 only
protocols (such as icmpv6) when it doesnt know we
have ipv6 addresses (link local).

So refreshing here is mandatory when we add/remove
our link-local addresses. As link-local addresses
are not added to ndb, so there is no change,
but ndb/cs still needs to get refreshed.