devip: don't add default broadcast addresses for ipv4 (thanks cgnarne)
When adding an ip address, we would also add a default
network mask (class A,B,C) in the selftab/routing table
based on the ip address only using defmask().

This can make some addresses unreachable when one uses
classless routing. So remove this, as the add command
already provides the exact netmask that should be used.